Update README.md
Welcome to the Plant Palette repository! This project is a modern, responsive website designed to showcase a variety of plants and provide users with a seamless browsing experience. From exploring different plant options to learning about plant care and contacting us for support, this website aims to be a comprehensive hub for plant enthusiasts.